test(aws/privesc): rename seeded test + merge live tests
Rename pkg/graph/queries/privesc_integration_test.go to
privesc_neo4j_seeded_test.go (package/build-tag unchanged). Merge the two recon
live tests into one pkg/modules/aws/recon/privesc_integration_test.go: the
pathfinding matrix is the survivor; the smoke test's 3 unique assertions are
folded in (suite-level multi-edge count, analysis-query path check; the
distinct-target <=10 bound is dropped as superseded by assertTargetAllowlist).

Relocate the kept synthetic helpers (SyntheticResource + Node, the kept-only
SyntheticComputeResources, SyntheticActionEdges, SameNodeStub*) into a new
test/testutil/privescsynth package so the live test and the replay test share
one source of truth, plus the typed-GAAD recon snapshot schema and an env-gated
capture hook (AURELIAN_CAPTURE_PRIVESC_SNAPSHOT=1, no-op otherwise). Internal
phase/bucket/method-NN comment jargon scrubbed in the same edit. No detection
code, terraform, or synthetic behavior changed.
